requirements.txt
BareNecessities==0.2.8
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mail==2.1.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
pycparser==2.20
requests==2.25.1
six==1.15.0
urllib3==1.26.5
Werkzeug==1.0.1